refactor: sync repositories with their aggregates
