+++ src/main/java/kr/co/wanted/backend31/repository/product/ProductCategoryRepository.java
@@ -1,4 +1,4 @@
-package kr.co.wanted.backend31.repository.category;
+package kr.co.wanted.backend31.repository.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
